requirements.txt
python-telegram-bot==22.7
python-dotenv==1.2.2
sentry-sdk==2.19.2
Flask==3.1.0

# Для разработки (опционально)
pytest==8.3.4
pytest-asyncio==0.25.2
